+++ lib/modules/datasource/github-tags/cache.ts
@@ -1,8 +1,6 @@
 import type { GithubHttp } from '../../../util/http/github';
-import {
-  AbstractGithubDatasourceCache,
-  StoredItemBase,
-} from '../github-releases/cache-base';
+import { AbstractGithubDatasourceCache } from '../github-releases/cache/cache-base';
+import type { StoredItemBase } from '../github-releases/cache/types';
 
 const query = `
 query ($owner: String!, $name: String!, $cursor: String, $count: Int!) {
